Refactor ContractsState into separate file
