+++ fuel-tx/src/transaction/types.rs
@@ -10,10 +10,7 @@ mod witness;
 pub use create::Create;
 pub use mint::Mint;
 pub use script::Script;
-pub use storage::{
-    ContractsStateData,
-    StorageSlot,
-};
+pub use storage::StorageSlot;
 pub use utxo_id::UtxoId;
 pub use witness::Witness;
