+++ fuel-vm/src/interpreter/blockchain/test/srwq.rs
@@ -2,13 +2,13 @@ use crate::{
     context::Context,
     storage::{
         ContractsState,
+        ContractsStateData,
         MemoryStorage,
     },
 };
 
 use super::*;
 use fuel_storage::StorageAsMut;
-use fuel_tx::ContractsStateData;
 use test_case::test_case;
 
 struct SRWQInput {
